Refactored intmodules to utils
